https://journals.openedition.org/transposition/5289?fbclid=IwAR2tnGDZoAMUsPin8vjfmDZ5abVknot4jZdxO8-uB8RD1fiyaw9jASgfmSg&lang=en
This essay explores the late engagement of music research with the long-standing yet overlooked association between music, violence and terror. Musicology and ethnomusicology have been historically and ideologically founded on the notion of music as an inherently positive, benign and ennobling artform. As such, these fields took long to engage with music’s long-standing association with violence, war and power, dating back to the Middle Ages if not earlier; the study of Nazi concentration and extermination camps is a notable exception, although it was also examined relatively belatedly. Left unexplored by historians, social anthropologists, human rights scholars, and legal experts working in the field of violence, detention and justice, this gap began to be critically addressed in music research only in the last decades. In mapping this new field, in this essay I explore (ethno)musicology’s latent reaction to music’s relationship with violence and technologies of terror as a disciplinary trauma, critically discussing a wide range of effects stemming from this silencing. Also examined is music’s integral role in new technologies of terror emerging during the Cold War, the cultural biases that have turned it into an elusive means of torture, and the effects stemming from the overshadowing of its damaging potential, as well as the practice of music torture, its impact and the legal context that has made it possible. Addressing the intersections of music and terror, I underline the need for more nuanced soundscapes of detention that explore the entanglement of negative and positive uses of music imposed from above and reclaimed from below.